Fix the bug of ids cache in case of multiple queries
There was indeed a bug, which occurred in the case of multiple queries for a single project.
Each query makes a request, receives IDs, and (this is the bug) overwrites the previous cache file.
So, the cache contains only a part of the actual total remote IDs, resulting in the reported bug (many local files missing their corresponding remote, so considered as obsolete).
The problem appears after the download because a new fetch is called and it uses the cache.

See line:
RCSB-Sync/src/project.py

Line 123 in 9a61a42

    remote_ids += search_and_download_ids(query, cache_file)

It doesn't make much sense for the search_and_download_ids function to save them in the cache as well.
Better to decouple (. In any case, it is an error in the function name.
This commit makes the decoupling.
Fix #2